more updates
typedefs for basic integer types and some other defines. also
implemented usage of these types across the code base

+++ string.h
@@ -5,12 +5,12 @@
 typedef struct
 {
 	char* buffer;
-	int length;
+	uint32 length;
 } String;
 
 extern String* String_Init(char* value);
 extern void String_Delete(String* string);
 
-extern int String_Length(char* value);
-extern int String_Compare(String* string1, String* string2);
-extern String* String_Substring(String* string, int startIndex, int count);
+extern uint32 String_Length(char* value);
+extern int16 String_Compare(String* string1, String* string2);
+extern String* String_Substring(String* string, uint32 startIndex, uint32 count);
